DJ/guvihack-main/backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
requests
livekit-api
